--- backend/.aws-sam/build.toml
@@ -1,12 +1,0 @@
-# This file is auto generated by SAM CLI build command
-
-[function_build_definitions.9e83863e-eb58-48c4-a510-e29e1097c1ff]
-codeuri = "/Users/stellalombrozo/Desktop/repos/samc-discounts/backend/built"
-runtime = "nodejs20.x"
-architecture = "x86_64"
-handler = "app.lambdaHandler"
-manifest_hash = ""
-packagetype = "Zip"
-functions = ["TypescriptFunction"]
-
-[layer_build_definitions]
