--- backend/built/jest.config.js
@@ -1,17 +1,0 @@
-"use strict";
-/*
- * For a detailed explanation regarding each configuration property and type check, visit:
- * https://jestjs.io/docs/configuration
- */
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.default = {
-    transform: {
-        '^.+\\.ts?$': 'ts-jest',
-    },
-    clearMocks: true,
-    collectCoverage: true,
-    coverageDirectory: 'coverage',
-    coverageProvider: 'v8',
-    testMatch: ['**/tests/unit/*.test.ts'],
-};
-//# sourceMappingURL=jest.config.js.map
