--- backend/discounts/src/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "hello_world",
-  "version": "1.0.0",
-  "description": "hello world sample for NodeJS",
-  "main": "app.js",
-  "repository": "https://github.com/awslabs/aws-sam-cli/tree/develop/samcli/local/init/templates/cookiecutter-aws-sam-hello-nodejs",
-  "author": "SAM CLI",
-  "license": "MIT",
-  "dependencies": {
-    "aws-sdk": "^2.1547.0",
-    "axios": ">=1.6.0"
-  },
-  "scripts": {
-    "test": "mocha tests/unit/",
-    "compile": "tsc"
-  },
-  "devDependencies": {
-    "@types/aws-lambda": "^8.10.133",
-    "@types/node": "^20.11.13",
-    "chai": "^4.3.6",
-    "mocha": "^10.2.0",
-    "typescript": "^5.3.3"
-  }
-}
