--- backend/package.json
@@ -1,14 +1,0 @@
-{
-  "dependencies": {
-    "aws-sdk": "^2.1540.0"
-  },
-  "scripts": {
-    "compile": "tsc",
-    "mock": "sam build --use-container && sam local invoke -e samc-discounts/events/event.json"
-  },
-  "devDependencies": {
-    "@types/aws-lambda": "^8.10.131",
-    "@types/node": "^20.11.5",
-    "typescript": "^5.3.3"
-  }
-}
