+++ backend/samc-be/events/event.json
@@ -1,5 +1,5 @@
 {
-  "body": "{\"message\": \"hello world\"}",
+  "body": "{\"name\": \"Bob\", \"surname\":\"Brown\",\"age\":21}",
   "resource": "/{proxy+}",
   "path": "/path/to/resource",
   "httpMethod": "POST",
