+++ backend/samc-be/samc-discounts/.eslintignore
