+++ backend/samc-be/samc-discounts/.eslintrc.js
