+++ backend/samc-be/samc-discounts/.npmignore
