+++ backend/samc-be/samc-discounts/.prettierrc.js
