+++ backend/samc-be/samconfig.toml
@@ -4,9 +4,10 @@ version = 0.1
 
 [default]
 [default.global.parameters]
-stack_name = "discounts"
+stack_name = "samc-be"
 
 [default.build.parameters]
+cached = true
 parallel = true
 
 [default.validate.parameters]
@@ -16,7 +17,6 @@ lint = true
 capabilities = "CAPABILITY_IAM"
 confirm_changeset = true
 resolve_s3 = true
-resolve_image_repos = true
 
 [default.package.parameters]
 resolve_s3 = true
