--- backend/samc-discounts/src/.npmignore
@@ -1,1 +1,0 @@
-tests/*
