--- backend/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "CommonJS",
-    "target": "ES2017",
-    "noImplicitAny": true,
-    "preserveConstEnums": true,
-    "outDir": "./built",
-    "sourceMap": true
-  },
-  "exclude": ["node_modules", "**/*.test.ts"]
-}
